requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
openai>=1.10.0
gradio>=4.15.0
streamlit>=1.29.0
pypdf>=3.17.0
chromadb>=0.4.20
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
deep-translator>=1.11.4
ragas>=0.1.1
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
tiktoken>=0.5.0
datasets>=2.16.0